requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# src/_github.py: 9
PyGithub == 1.35

# src/_mongodb.py: 6
pymongo == 3.4.0
